fix(ui): initial success animation on import wallet (#2881)
Description
---

if the imported wallet had any `unpaid` xtm in pool stats the initial
change for `rewardValue` would trigger the success animation. this,
coupled with the fact that mining stops on import, and does not start up
again because of the initiated checks (will look at that in a separate
PR) caused the tower to get stuck on success and not move on to the next
(non-existent) state

- reset the pool mining store to initial state when importing seed words
- made the first diff check more strict by adding a check for is mining
- passed the `isReplay` flag to `setAnimationState` triggered by the
pool reward _juuust in case_ the reward animation is triggered mid
mining stop, to avoid it getting stuck

Motivation and Context
---
- fixes https://github.com/tari-project/universe/issues/2759

How Has This Been Tested?
---

- locally


https://github.com/user-attachments/assets/e77500e2-771a-4892-a50f-2a527de460fc

---------

Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>