fix: tx history list (#2906)
Description
---

for #2874:
- add character limit (240c) for tx inputs


for #2713: 

the main issue is that we don't re-can and emit wallet scanning progress
if the initial scan is already done, BUT the wallet phase _is_ restarted
(when switching wallet types, or node types), and sometimes the
healthcheck returns unhealthy (not started) so we don't get any txs back
until the walelt scans back up

- added a new event with loading state (and `unhealthy` for future use
if needed - might remove?) and used where initial scan is skipped so the
FE knows to put the wallet into the loading/scanning state
- cleanups - removed the old List Loading UI which we don't get to
anymore
- cleaned up `useFetchTxHistory` and the list component


Motivation and Context
---
- closes https://github.com/tari-project/universe/issues/2874
- closes https://github.com/tari-project/universe/issues/2713

How Has This Been Tested?
---


https://github.com/user-attachments/assets/912925fa-92bc-47ca-b90d-594177794b29


What process can a PR reviewer use to test or verify this change?
---

- load up wallet and view your TX history
- switch node types (best to check from `Local` to `Remote` and/or add
your address to mine directly to an exchange, and then back to TU

---------

Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>