fix(ui): mode selector close (#3055)
Description
---

- fixed the `MiningMode` open/close - made a function `handleOpenChange`
to be used post selecting the mode as well as for the `useFloating`
hook. included clearing the listRef in there (so the selected mode after
changing will always be updated)
  - also moved the `AnimatePresence` to just above the `List` 
- fixed the eslint config (it wasn't picking up the exhaustive deps
rules)
  - fixed the lints that were now coming up 


Motivation and Context
---
- fixes #3052 

How Has This Been Tested?
---
- locally:


https://github.com/user-attachments/assets/d06db2ad-cca7-4ec7-8806-a24807f0fc22


What process can a PR reviewer use to test or verify this change?
---
- should be mining
- switch through the modes

Breaking Changes
---

- [x] None

Co-authored-by: Brian Pearce <brianp@users.noreply.github.com>